Fix workspace clippy usage

+++ framework/libs/linux-bzimage/setup/src/main.rs
@@ -40,5 +40,5 @@ fn get_payload(boot_params: &BootParams) -> &'static [u8] {
     let payload_length = hdr.payload_length as usize;
     // Safety: the payload_offset and payload_length is valid if we assume that the
     // boot_params struct is correct.
-    unsafe { core::slice::from_raw_parts_mut(payload_offset as *mut u8, payload_length as usize) }
+    unsafe { core::slice::from_raw_parts_mut(payload_offset as *mut u8, payload_length) }
 }
