+++ framework/libs/linux-bzimage/setup/src/x86/amd64_efi/mod.rs
@@ -13,8 +13,8 @@ global_asm!(include_str!("setup.S"));
 pub const ASTER_ENTRY_POINT: u32 = 0x8001200;
 
 unsafe fn call_aster_entrypoint(entrypoint: u64, boot_params_ptr: u64) -> ! {
-    asm!("mov rsi, {}", in(reg) boot_params_ptr as u64);
-    asm!("mov rax, {}", in(reg) entrypoint as u64);
+    asm!("mov rsi, {}", in(reg) boot_params_ptr);
+    asm!("mov rax, {}", in(reg) entrypoint);
     asm!("jmp rax");
 
     unreachable!();
